+++ test/worker_test.go
@@ -44,7 +44,7 @@ func (w *TestWorker) Work(ctx context.Context, request *pbsubstreams.Request, _ 
 	ctx = reqctx.WithLogger(ctx, logger)
 
 	logger.Info("worker running job",
-		zap.String("output_module", request.GetOutputModuleName()),
+		zap.String("output_module", request.MustGetOutputModuleName()),
 		zap.Int64("start_block_num", request.StartBlockNum),
 		zap.Uint64("stop_block_num", request.StopBlockNum),
 	)
@@ -55,7 +55,7 @@ func (w *TestWorker) Work(ctx context.Context, request *pbsubstreams.Request, _ 
 		}
 	}
 	logger.Info("worker done running job",
-		zap.String("output_module", request.GetOutputModuleName()),
+		zap.String("output_module", request.MustGetOutputModuleName()),
 		zap.Int64("start_block_num", request.StartBlockNum),
 		zap.Uint64("stop_block_num", request.StopBlockNum),
 	)